+++ pom.xml
@@ -61,6 +61,12 @@
             <artifactId>azure-data-tables</artifactId>
             <version>12.1.1</version>
         </dependency>
+    
+      <dependency>
+        <groupId>com.fasterxml.jackson.core</groupId>
+        <artifactId>jackson-databind</artifactId>
+        <version>2.10.0.pr1</version>
+      </dependency>
     </dependencies>
 
     <build>
